% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_resd_streaming_algorithm.R
\name{auto_resd_streaming_algorithm}
\alias{auto_resd_streaming_algorithm}
\title{Function to automatically tune window sizes if there's an anomaly in the learning window}
\usage{
auto_resd_streaming_algorithm(
  xdata,
  anomalies_known,
  alpha,
  wprime,
  w,
  iters,
  k
)
}
\arguments{
\item{xdata}{Raw_data[1:1055,2]}

\item{anomalies_known}{known anomaly labels}

\item{alpha}{Significance level in t distribution critical value.}

\item{wprime}{Initialisation window size, wprime>w e.g. wprime=100.}

\item{w}{Integer sliding window size  e.g.w=50.}

\item{iters}{numeric integer number of iterations.}

\item{k}{maximum number of anomalies in a given window, e.g.  k=10 by definition of anomaly k<=0.5(length(xdata)).}
}
\value{
anomaly_results list of the following.

anomaly_results$x_A the ith row of this matrix is the anomalies found using the Sequential Grubbs algorithm in the ith window

anomaly_results$t_A  the time stamp of the detected anomaly LATER

anomaly_results$t_window the time stamp of the last observation in the relevant window
}
\description{
Function to automatically tune window sizes if there's an anomaly in the learning window
}
