% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esd_grubbs_test.R
\name{esd_grubbs_test}
\alias{esd_grubbs_test}
\title{Extreme Studentised Deviate test for up to k anomalies in a vector of residuals
Recursive updates are enabled by expressing the ESD test statistic as function of the Grubbs ratio (Grubbs 1950)}
\usage{
esd_grubbs_test(e0bar, epsilon, k, alpha, ss, w)
}
\arguments{
\item{e0bar}{Mean of the vector of time series residuals.}

\item{epsilon}{Time series residuals.}

\item{k}{Number of outliers to test.}

\item{alpha}{Significance level.}

\item{ss}{Initial sums of squares.}

\item{w}{Size of window (This is denoted n in the Rosner paper)}
}
\value{
x_A Anomaly vector.

grubbs_results List

grubbs_results$x_A The anomalies found using the Sequential Grubbs test

grubbs_results$t_A  the time stamp or index of the detected anomaly LATER
}
\description{
Extreme Studentised Deviate test for up to k anomalies in a vector of residuals
Recursive updates are enabled by expressing the ESD test statistic as function of the Grubbs ratio (Grubbs 1950)
}
