% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{model_fitting}
\alias{model_fitting}
\title{Model fitting functions used in to find a useful model based on the initialisation phase}
\usage{
model_fitting(xprime, timeprime)
}
\arguments{
\item{xprime}{time series initialisation dataset as a vector.}

\item{timeprime}{time vector (for now must be regular)}
}
\value{
streg fitted model
}
\description{
Model fitting functions used in to find a useful model based on the initialisation phase
}
