% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodogram.R
\name{periodogram}
\alias{periodogram}
\title{Taken from tsa which had depreciated and was taken off cran.}
\usage{
periodogram(
  y,
  log = "no",
  plot = TRUE,
  ylab = "Periodogram",
  xlab = "Frequency",
  lwd = 2,
  ...
)
}
\description{
Taken from tsa which had depreciated and was taken off cran.
}
