% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resd_initialisation.R
\name{resd_initialisation}
\alias{resd_initialisation}
\title{Model fitting functions used in to find an model based on the initialisation phase}
\usage{
resd_initialisation(xdata, wprime, w, tt)
}
\arguments{
\item{xdata}{Raw_data[1:1055,2]}

\item{wprime}{Initialisation window size, wprime>w e.g. wprime=100.}

\item{w}{Integer sliding window size  e.g.w=50.}

\item{tt}{time to start streaming tt=wprime here or w but could be any time of interest.}
}
\value{
initial fitted model a list of the following
}
\description{
Model fitting functions used in to find an model based on the initialisation phase
}
