% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resd_initialisation.R
\name{resd_initialisation}
\alias{resd_initialisation}
\title{Functions used to find a useful model forecast based on the initialisation phase / wprime window}
\usage{
resd_initialisation(xdata, wprime, w)
}
\arguments{
\item{xdata}{Univariate time series dataset}

\item{wprime}{Initialisation window size}

\item{w}{Integer sliding window size, w<wprime}
}
\value{
initial fitted model a list of the following
}
\description{
Functions used to find a useful model forecast based on the initialisation phase / wprime window
}
